+++ src/AppProvider.tsx
@@ -17,6 +17,7 @@ import { I18nProvider } from "./providers/I18nProvider/I18nProvider";
 import { CatalogProvider } from "./providers/CatalogProvider/CatalogProvider";
 import { QuotesProvider } from "./providers/QuotesProvider/QuotesProvider";
 import { NotificationProvider } from "./providers/AlertProvider/NotificationProvider";
+import { StepsProvider } from "./providers/StepsProvider/StepsProvider";
 
 export const AppProvider: React.FC<PropsWithChildren<WidgetConfigProps>> = ({
   children,
@@ -30,13 +31,15 @@ export const AppProvider: React.FC<PropsWithChildren<WidgetConfigProps>> = ({
           <ThemeProvider>
             <WalletProvider>
               <StoreProvider config={config} formRef={formRef}>
-                <CatalogProvider>
-                  <NotificationProvider>
-                    <QuotesProvider>
-                      <AppRouter>{children}</AppRouter>
-                    </QuotesProvider>
-                  </NotificationProvider>
-                </CatalogProvider>
+                <AppRouter>
+                  <CatalogProvider>
+                    <NotificationProvider>
+                      <QuotesProvider>
+                        <StepsProvider>{children}</StepsProvider>
+                      </QuotesProvider>
+                    </NotificationProvider>
+                  </CatalogProvider>
+                </AppRouter>
               </StoreProvider>
             </WalletProvider>
           </ThemeProvider>