+++ src/components/TokenList/TokenList.tsx
@@ -3,8 +3,6 @@ import { Box } from "@mui/material";
 import { type FC } from "react";
 import { useChain } from "../../hooks/useChain.js";
 import { useDebouncedWatch } from "../../hooks/useDebouncedWatch";
-import { FormKeyHelper } from "../../stores/form/types.js";
-import { useFieldValues } from "../../stores/form/useFieldValues.js";
 import { TokenNotFound } from "./TokenNotFound.js";
 import { VirtualizedTokenList } from "./VirtualizedTokenList.js";
 import type { TokenListProps } from "./types.js";
@@ -17,17 +15,14 @@ export const TokenList: FC<TokenListProps> = ({
   height,
   onClick,
 }) => {
-  const [selectedChainId] = useFieldValues(FormKeyHelper.getChainKey(formType));
   const [tokenSearchFilter]: string[] = useDebouncedWatch(
     320,
     "tokenSearchFilter"
   );
 
-  const { chain: selectedChain } = useChain(selectedChainId);
+  const { account } = useAccount();
+  const { chain: selectedChain } = useChain(account?.chainId);
 
-  const { account } = useAccount({
-    chainType: selectedChain?.network_type,
-  });
   const { balances, isLoading } = useTokenBalances(
     account?.address ?? "",
     selectedChain ?? undefined
@@ -45,7 +40,7 @@ export const TokenList: FC<TokenListProps> = ({
         account={account}
         tokens={balances}
         scrollElementRef={parentRef}
-        chainId={selectedChainId}
+        chainId={account?.chainId}
         chain={selectedChain}
         isLoading={isLoading}
         showCategories={showCategories}