+++ src/hooks/useTransactionFlow.ts
@@ -9,12 +9,15 @@ import { useTransactionHelpers } from "./useTransactionHelpers";
 import { useFetch } from "./useFetch";
 import { useToken } from "./useToken";
 import { useNotificationContext } from "../providers/AlertProvider/NotificationProvider";
+import { useSteps } from "../providers/StepsProvider/StepsProvider";
+import { useCallback } from "react";
 
 export const useTransactionFlow = () => {
   const navigate = useNavigate();
   const { product } = useProduct();
   const tokenKey = FormKeyHelper.getTokenKey("from");
   const { quote } = useQuotes();
+  const { clearStep } = useSteps();
   const [chainId, tokenAddress, reference, requiredFields] = useFieldValues(
     FormKeyHelper.getChainKey("from"),
     tokenKey,
@@ -42,9 +45,10 @@ export const useTransactionFlow = () => {
               product,
               token,
             });
-
+            clearStep();
             navigate(`/status/${data?.transaction_intent?.id}`, { state: { signature } });
           } catch (error) {
+            clearStep();
             showNotification("error", "Error handling the transaction signature");
             console.error("Error handling the transaction signature:", error);
           }
@@ -56,7 +60,7 @@ export const useTransactionFlow = () => {
     },
   });
 
-  const handleTransaction = async () => {
+  const handleTransaction = useCallback(() => {
     mutate({
       reference: reference,
       required_fields: requiredFields,
@@ -69,7 +73,7 @@ export const useTransactionFlow = () => {
         wallet: account?.address,
       },
     });
-  };
+  }, [mutate, reference, requiredFields, product?.sku, chain?.key, quote?.digital_asset, quote?.digital_asset_amount, account?.address]);
 
   return {
     handleTransaction,